# Conflicts:
#	app/jobs/zip_pdf_details_job.rb

+++ lib/tasks/boats_zip.rake
@@ -1,8 +1,12 @@
 namespace :boats_zips do
   desc 'Delete old dirs inside zipped_pdf dir'
+  week_ago = 1.week.ago
+
   task :cleanup do
     Dir['public/zipped_pdfs/*'].each do |file|
-      FileUtils.rm_rf(file)
+      if File.mtime(file) < week_ago
+        FileUtils.rm_rf(file)
+      end
     end
   end
 end
